Fix minor bug and compile time warnings
git-svn-id: https://lmms.svn.sf.net/svnroot/lmms/trunk/lmms@1082 0778d3d1-df1d-0410-868b-ea421aaaa00d

+++ include/effect_lib.h
@@ -448,7 +448,7 @@ namespace effectLib
 		// Lou's Hack
 		void setWideCoeff( const float _wideCoeff )
 		{
-			m_wideCoeff = _wideCoeff;
+			m_wideCoeff = static_cast<int>( _wideCoeff );
 		}
 		
 		int getWideCoeff()
